requirements.txt
alembic==1.4.0
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
boto3==1.12.26
botocore==1.15.26
cffi==1.14.0
Click==7.0
distlib==0.3.0
dnspython==1.16.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
pbr==5.4.4
protobuf==3.6.1
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.13
stevedore==1.32.0
urllib3==1.25.8
virtualenv==20.0.4
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.16.1
wrapt==1.11.2
zipp==3.0.0
